flask-app/requirements.txt
Flask==1.1.2
flask_oidc==1.4.0
oauth2client==4.1.3
authlib==0.15.4
requests==2.26.0
flask_login==0.5.0
Flask-Migrate==3.1.0
flask_sqlalchemy==2.5.1
flask_wtf==0.15.1
Jinja2==3.0.1
mgrs==1.4.0
numpy==1.21.1
pandas==1.3.1
werkzeug==2.0.1
WTForms==2.3.3
xlrd==2.0.1
openpyxl==3.0.7
psycopg2==2.9.1
pyjwt==2.1.0
redis==3.5.3
rq==1.9.0
boto3==1.18.16
